Remove unused binary search and bubble sort implementations
Former-commit-id: b47526f1219405f464a4614955594084439000f5 [formerly 7ffda05058187185652d45565ec0adbb00eb023b]
Former-commit-id: bd70e72f2c9ff7a1b1a2bd491146c69f06acbd24
Former-commit-id: 384f4a4e6fc5e679a128d61283d465ae68cec1bb

+++ sem4/daa/bblSort.c
@@ -4,11 +4,10 @@
 
 int main(){
     int i, j,n=4;
-    float arr[20], temp;
-//  int arr[20] = {4,3,8,1, 9};
+    int arr[20], temp;
     for(i=0; i<n; i++){
         printf("Enter Element : ");
-        scanf("%f", &arr[i]);
+        scanf("%d", &arr[i]);
     }
     for(i=0; i<n; i++){
         for(j=0; j<n-i-1; j++){
@@ -20,7 +19,7 @@ int main(){
         }
     }
     for(i=0; i<n; i++){
-        printf("%f, ", arr[i]);
+        printf("%d, ", arr[i]);
     }
 }
 // 9 7 6 3 1
